Reorganized
The navigation bar has now been reorganized. Projects, designs, gallery, and Knåker have now been placed under the label "Works," a new page with nice icons that links to these aforementioned pages.
Removed and added a few pictures to the gallery.
Edited the text on some of the pages.